scx_wd40: fix build and build in CI
libbpf-cargo was updated in #2397. This was very invasive and I missed
wd40 while doing this because `cargo build` & the CI don't build wd40.

Fix the build, and add wd40 to the Cargo workspace and therefore the CI
builds. This doesn't test wd40 but does build it. This will prevent
silly regressions like this in the future.

Test plan:
- CI